requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-pptx==0.6.23
openpyxl==3.1.2
Pillow==10.0.1
speechrecognition==3.10.0
pydub==0.25.1
opencv-python==4.8.1.78
python-docx==0.8.11
reportlab==4.0.4
flask-socketio==5.3.6
python-socketio==5.9.0
eventlet==0.33.3
